+++ app/privacy.tsx
@@ -1,0 +1,5 @@
+import { PrivacyNoticeScreen } from "../src/ui/screens/PrivacyNoticeScreen";
+
+export default function PrivacyRoute() {
+  return <PrivacyNoticeScreen />;
+}